+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @talon-cli/cli
 
+## 1.1.0
+
+### Minor Changes
+
+- create 包支持模板下载
+
+### Patch Changes
+
+- Updated dependencies
+  - @talon-cli/create@1.1.0
+
+## 1.0.2
+
+### Patch Changes
+
+- @talon-cli/create@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes