+++ packages/create/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @talon-cli/create
 
+## 1.1.0
+
+### Minor Changes
+
+- create 包支持模板下载
+
+### Patch Changes
+
+- Updated dependencies
+  - @talon-cli/utils@1.2.0
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @talon-cli/utils@1.1.1
+
 ## 1.0.1
 
 ### Patch Changes