+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @talon-cli/utils
 
+## 1.2.0
+
+### Minor Changes
+
+- create 包支持模板下载
+
+## 1.1.1
+
+### Patch Changes
+
+- utils 包增加统一导出入口
+
 ## 1.1.0
 
 ### Minor Changes